html/semantics/scripting-1/the-script-element/scriptexecute-events/004.html

<!DOCTYPE html>
<html>
  <title>no beforescriptexecute on unsupported script type</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    document.addEventListener("beforescriptexecute", function(e) {targets.push(e.target);}, false);
    var t = async_test();
 </script>
 <script type="text/plain">
    t.step(function() {assert_unreached()});
 </script>
  <script>
   t.step(function() {assert_array_equals(targets, [document.getElementsByTagName("script")[4]]);})
   t.done();
 </script>
